refactor: move test to another pkg
